+++ client/src/components/header/Header.tsx
@@ -1,7 +1,7 @@
 import { Box, Flex, Heading, HStack, Text } from "@chakra-ui/react";
 import NextLink from "next/link";
 import React from "react";
-import UserMenu from "./user-menu";
+import UserMenu from "./UserMenu";
 
 type Props = {
   contentTitle?: string;
